Refactor IP utilities to allow for better reuse of them.
Move IP utilities from probes/probeutils to common/iputils.

I found myself using the same methods in the file_based targets implementation.

PiperOrigin-RevId: 317436181
